+++ Sources/Requests/Requests+SessionUpdateRequest1.swift
@@ -6,6 +6,11 @@ extension Requests {
         public let comments: String?
         public let deleteComments: Bool?
         public let `public`: Bool?
+        public let offChrtOrderId: String?
+        public let deleteOffChrtOrderId: Bool?
+        public let flightNumber: String?
+        public let deleteFlightNumber: Bool?
+        public let faFlightIds: [String]?
         /// Additional properties that are not explicitly defined in the schema
         public let additionalProperties: [String: JSONValue]
 
@@ -14,12 +19,22 @@ extension Requests {
             comments: String? = nil,
             deleteComments: Bool? = nil,
             public: Bool? = nil,
+            offChrtOrderId: String? = nil,
+            deleteOffChrtOrderId: Bool? = nil,
+            flightNumber: String? = nil,
+            deleteFlightNumber: Bool? = nil,
+            faFlightIds: [String]? = nil,
             additionalProperties: [String: JSONValue] = .init()
         ) {
             self.sessionId = sessionId
             self.comments = comments
             self.deleteComments = deleteComments
             self.public = `public`
+            self.offChrtOrderId = offChrtOrderId
+            self.deleteOffChrtOrderId = deleteOffChrtOrderId
+            self.flightNumber = flightNumber
+            self.deleteFlightNumber = deleteFlightNumber
+            self.faFlightIds = faFlightIds
             self.additionalProperties = additionalProperties
         }
 
@@ -29,6 +44,11 @@ extension Requests {
             self.comments = try container.decodeIfPresent(String.self, forKey: .comments)
             self.deleteComments = try container.decodeIfPresent(Bool.self, forKey: .deleteComments)
             self.public = try container.decodeIfPresent(Bool.self, forKey: .public)
+            self.offChrtOrderId = try container.decodeIfPresent(String.self, forKey: .offChrtOrderId)
+            self.deleteOffChrtOrderId = try container.decodeIfPresent(Bool.self, forKey: .deleteOffChrtOrderId)
+            self.flightNumber = try container.decodeIfPresent(String.self, forKey: .flightNumber)
+            self.deleteFlightNumber = try container.decodeIfPresent(Bool.self, forKey: .deleteFlightNumber)
+            self.faFlightIds = try container.decodeIfPresent([String].self, forKey: .faFlightIds)
             self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
         }
 
@@ -39,6 +59,11 @@ extension Requests {
             try container.encodeIfPresent(self.comments, forKey: .comments)
             try container.encodeIfPresent(self.deleteComments, forKey: .deleteComments)
             try container.encodeIfPresent(self.public, forKey: .public)
+            try container.encodeIfPresent(self.offChrtOrderId, forKey: .offChrtOrderId)
+            try container.encodeIfPresent(self.deleteOffChrtOrderId, forKey: .deleteOffChrtOrderId)
+            try container.encodeIfPresent(self.flightNumber, forKey: .flightNumber)
+            try container.encodeIfPresent(self.deleteFlightNumber, forKey: .deleteFlightNumber)
+            try container.encodeIfPresent(self.faFlightIds, forKey: .faFlightIds)
         }
 
         /// Keys for encoding/decoding struct properties.
@@ -47,6 +72,11 @@ extension Requests {
             case comments
             case deleteComments = "delete_comments"
             case `public`
+            case offChrtOrderId = "off_chrt_order_id"
+            case deleteOffChrtOrderId = "delete_off_chrt_order_id"
+            case flightNumber = "flight_number"
+            case deleteFlightNumber = "delete_flight_number"
+            case faFlightIds = "fa_flight_ids"
         }
     }
 }